requirements.txt
crewai==0.175.0
fastapi==0.116.1
joblib==1.5.2
numpy==2.3.2
pandas==2.3.2
pydantic==2.11.7
python-dotenv==1.1.1
Requests==2.32.5
scikit_learn==1.7.1
streamlit==1.49.1
